force_cast (11 errors) shorthand_operator (10 errors) large_tuple (2 errors, simply disabled in comment line) file_length (1 error, 7 warnings, 1 class is refactored, the others are disabled) type_body_length (1 error, 8 warnings)
- No swiftlint errors (serious violation) remaining so far.
